The Full Game
Added: The computer may now choose a color to go first, unlike the previous versions.
Added: Clicking on a white square triggers a clicking sound, and the cursor changes to make it more clear for the user where they should click.
Added: The "rain" of the winning animation now has two colors of "raindrops" according to the winner.
Fixed: the entire row of raindrops would spawn in the middle of the screen first and only after few second relocated to a different location.
Fixed: raindrops will no-longer appear on the edge of the screen.
Upgraded: the start animation is now better as the compute takes few milliseconds more before marking the first to go player.